ajout set location depuis adresse récupérer du formulaire, migration sur base sql lite et refactoring code

+++ src/Controller/AdminController.php
@@ -32,14 +32,32 @@ class AdminController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $address = $form->getData()->getAddress();
+            $search_url = "https://nominatim.openstreetmap.org/search?q=.$address.&format=json";
+            $httpOptions = [
+                "http" => [
+                    "method" => "GET",
+                    "header" => "User-Agent: Nominatim-Test"
+                ]
+            ];
+            $streamContext = stream_context_create($httpOptions);
+            $json = file_get_contents($search_url, false, $streamContext);
+            
+            $decoded = json_decode($json, true);
+            $lat = floatval($decoded[0]["lat"]);
+            $lon = floatval($decoded[0]["lon"]);
+            $agency->setLat($lat);
+            $agency->setLon($lon);
             $em->persist($agency);
             $em->flush();
-
+            
+            
             $this->addFlash('success', 'Agence ajouter avec succès !');
-
+            
             return $this->redirectToRoute('app_admin');
         }
-
+        
+    
         return $this->render('admin/edit.html.twig', [
             'form' => $form->createView(),
         ]);
@@ -53,6 +71,22 @@ class AdminController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $address = $form->getData()->getAddress();
+            $search_url = "https://nominatim.openstreetmap.org/search?q=.$address.&format=json";
+            $httpOptions = [
+                "http" => [
+                    "method" => "GET",
+                    "header" => "User-Agent: Nominatim-Test"
+                ]
+            ];
+            $streamContext = stream_context_create($httpOptions);
+            $json = file_get_contents($search_url, false, $streamContext);
+            
+            $decoded = json_decode($json, true);
+            $lat = floatval($decoded[0]["lat"]);
+            $lon = floatval($decoded[0]["lon"]);
+            $agency->setLat($lat);
+            $agency->setLon($lon);
             $em->flush();
 
             $this->addFlash('success', 'Agence modifier avec succès !');
@@ -65,14 +99,16 @@ class AdminController extends AbstractController
         ]);
     }
     /**
-     * @Route("/admin/map_delete/{id<[0-9]+>}", name="app_map_delete" )
+     * @Route("/admin/map_delete/{id<[0-9]+>}", name="app_map_delete", methods={"DELETE"})
      */
     public function mapDelete(Request $request, Agency $agency, EntityManagerInterface $em)
     {
+        if ($this->isCsrfTokenValid('agency_deletion_' . $agency->getId(), $request->request->get('csrf_token'))) {
             $em->remove($agency);
             $em->flush();
 
             $this->addFlash('info', 'Centre supprimer avec succès !');
+        }
         return $this->redirectToRoute('app_admin');
     }
 }
